removed dependency on 'highp' shader precision on devices without 'highp' support

+++ src/shader/blur.fs
@@ -1,6 +1,5 @@
 #ifdef GL_ES
-  // we need high precision for the depth values
-  precision highp float;
+  precision mediump float;
 #endif
 
 uniform sampler2D uTexIndex;
